Updated code points to geom v0.1.0.
Updated to use slippy tile from geom v0.1.0.
This fixes the issues with slippy tile.

+++ provider/gpkg/gpkg.go
@@ -117,11 +117,11 @@ func (p *Provider) TileFeatures(ctx context.Context, layer string, tile provider
 		qtext = fmt.Sprintf("%v FROM `%v` l JOIN `%v` si ON l.`%v` = si.id WHERE l.`%v` IS NOT NULL AND !BBOX! ORDER BY l.`%v`", selectClause, pLayer.tablename, rtreeTablename, pLayer.idFieldname, pLayer.geomFieldname, pLayer.idFieldname)
 
 		z, _, _ := tile.ZXY()
-		qtext = replaceTokens(qtext, z, tileBBox)
+		qtext = replaceTokens(qtext, uint(z), tileBBox)
 	} else {
 		// If layer was specified via "sql" in config, collect it
 		z, _, _ := tile.ZXY()
-		qtext = replaceTokens(pLayer.sql, z, tileBBox)
+		qtext = replaceTokens(pLayer.sql, uint(z), tileBBox)
 		qtext = queryParams.ReplaceParams(qtext, &args)
 	}
